fix: 修复上传图片时 size 字段 JSON 解析失败
API 返回的 size 字段是浮点数，但代码定义为 int64，导致解析失败。
将 Size 字段类型从 int64 改为 float64。

+++ internal/api/images.go
@@ -13,7 +13,7 @@ type UploadResponse struct {
 	Data    struct {
 		Key        string `json:"key"`
 		OriginName string `json:"origin_name"`
-		Size       int64  `json:"size"`
+		Size       float64 `json:"size"`
 		Mimetype   string `json:"mimetype"`
 		Width      int    `json:"width"`
 		Height     int    `json:"height"`
@@ -54,7 +54,7 @@ type ImagesListResponse struct {
 		Images      []struct {
 			Key        string `json:"key"`
 			OriginName string `json:"origin_name"`
-			Size       int64  `json:"size"`
+			Size       float64 `json:"size"`
 			Mimetype   string `json:"mimetype"`
 			Width      int    `json:"width"`
 			Height     int    `json:"height"`
